Move get_check_function to caller #2
CMK-12298
