+++ tests/unit/cmk/base/agent_based/checking/test_checking.py
@@ -18,6 +18,7 @@ from cmk.checkengine.legacy import LegacyCheckParameters
 from cmk.checkengine.parameters import TimespecificParameters, TimespecificParameterSet
 
 import cmk.base.agent_based.checking as checking
+import cmk.base.checkers as checkers
 import cmk.base.config as config
 from cmk.base.api.agent_based.checking_classes import consume_check_results, Metric, Result, State
 from cmk.base.plugins.agent_based.agent_based_api.v1.type_defs import CheckResult
@@ -225,7 +226,7 @@ def test_time_resolved_check_parameters(
     ],
 )
 def test_aggregate_result(subresults: CheckResult, aggregated_results: ServiceCheckResult) -> None:
-    assert checking._aggregate_results(consume_check_results(subresults)) == aggregated_results
+    assert checkers._aggregate_results(consume_check_results(subresults)) == aggregated_results
 
 
 def test_config_cache_get_clustered_service_node_keys_no_cluster(monkeypatch: MonkeyPatch) -> None:
